requirements.txt
Markdown==3.7
python-dotenv==1.0.1
types-Markdown==3.7.0.20241204
